Deterministic event loop in FE (#577)
A sequential strictly ordered input/output event stream is implemented.
The event loop generate output events based on user code activity. The
event loop output events a going in strict deterministic order if user
code is deterministic.

The event loop applies its input event (like function call completion
events) in a strict deterministic order to user code.

Also deduped application deployments in application tests because they
otherwise makes the tests running very long which is wasting time.

I did a large internal refactoring here that simplifies handling of map
reduce operations which resulted in each map and reduce operation always
having a function call that represents their results. This also allows
now to return map operations as tail calls and pass a future as map and
reduce operation inputs. The cost of this simplification and extra
features is 2 extra function call per map operation and 1 extra function
call per reduce operation. This simplification though unlocks adding
more complexity now for durable execution and other future features like
durable timers and suspend.

map-reduce benchmark performance didn't change:

```
poetry run python ../indexify/benchmarks/map_reduce/main.py --num-requests 5 --maps-count 100
main:

Deploying benchmark application...
Starting map reduce benchmark with maps_count: 100, num_requests: 5
Started request faFZ-7Yy8UK8davhcd3I4 with 100 functions
Started request bgK3ozmp5CTM0lm8iUUSB with 100 functions
Started request b2SYCL-CApyf79bwSl9bi with 100 functions
Started request M8i64hiiwAavBoLaQt6O5 with 100 functions
Started request L2yxCqtiE8Ky3VBsUtzq- with 100 functions
All 5 requests started in 0.01 seconds
Waiting for all requests to complete...

Completed 5 requests in 64.04s

Test Configuration:
- Requests: 5
- Map calls per request: 100
- Total function calls: 1000

PR:

Deploying benchmark application...
Starting map reduce benchmark with maps_count: 100, num_requests: 5
Started request d8p2LBcqjuTdkmOso3NLc with 100 functions
Started request esMWNI9z1dCGprk8eugNy with 100 functions
Started request T-qQPZSRTlv3TvfAjSVp7 with 100 functions
Started request ToKpJjIudUsJ9Z4F9wn0r with 100 functions
Started request ETyaz6oUaFKRXrfFOPqBS with 100 functions
All 5 requests started in 0.01 seconds
Waiting for all requests to complete...

Completed 5 requests in 66.68s

Test Configuration:
- Requests: 5
- Map calls per request: 100
- Total function calls: 1000
```

+++ tests/run_tests.sh
@@ -7,13 +7,6 @@ if [[ -z "$TENSORLAKE_API_URL" ]]; then
     exit 1
 fi
 
-echo "Building Rust Cloud SDK Python bindings..."
-poetry run maturin develop --manifest-path ../crates/rust-cloud-sdk-py/Cargo.toml
-if [ $? -ne 0 ]; then
-  echo "Failed to build Rust Cloud SDK Python bindings." 1>&2
-  exit 1
-fi
-
 enable_fe_test_suite=false
 enable_applications_test_suite=false
 enable_cli_test_suite=false
@@ -56,6 +49,13 @@ run_test_suite() {
 
 # cd to the script's directory.
 cd "$(dirname "$0")"
+
+echo "Building Rust Cloud SDK Python bindings..."
+poetry run maturin develop --manifest-path ../crates/rust-cloud-sdk-py/Cargo.toml
+if [ $? -ne 0 ]; then
+  echo "Failed to build Rust Cloud SDK Python bindings." 1>&2
+  exit 1
+fi
 
 summary_file=".run_tests_summary.txt"
 rm -f $summary_file
